fix: expose parent directory in device-config plug
This is required to allow providing a whole directory containing configuration files and other directories.

Related to canonical/edgex-config-provider#2

In addition, the redundant content identifier is removed because it is equal to the plug name. When not set, it default to the plug name; see https://snapcraft.io/docs/content-interface

Signed-off-by: Farshid Tavakolizadeh <farshid.tavakolizadeh@canonical.com>